+++ FinerioConnectCore/src/main/java/com/finerioconnect/core/sdk/models/FCResumeByMonth.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class FCResumeByMonth(
-    @Json(name = "date") val date : Int,
+    @Json(name = "date") val date : Long,
     @Json(name = "amount") val amount : Double,
     @Json(name = "categories") val categories : List<FCResumeByCategory>
 )
